standalone-healthcare-app/backend/pyproject.toml
[tool.poetry]
name = "healthcare-backend"
version = "0.1.0"
description = "FastAPI backend for healthcare chat application"
authors = ["Devin AI <devin-ai-integration[bot]@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110.0"
uvicorn = ">=0.30.0"
anyio = ">=4.5"
pydantic = "^2.6.3"
openai = "^1.25.0"
python-dotenv = "^1.0.1"
starlette = "^0.36.3"
fastapi-cors = "^0.0.6"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
